requirements.txt
python-telegram-bot ~= 13.8.1